Completely refactored the update request functionality so that no hash maps are used whatsoever. Duplicate request functionality works perfectly (as far as I can tell), and I also added the ability to see the history of the user's update requests.
With this functionality now in place, maybe the timestamp could be added so that the user can re-submit a request within a week or so. (This would be as simple as adding another field to the SoftwarePoint class and keeping track of it for each object in the file). This is up to the stakeholders however...